% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_labels.R
\name{arial}
\alias{arial}
\title{Change plotly font to Arial}
\usage{
arial(p, size = 14)
}
\arguments{
\item{p}{A plotly object}

\item{size}{Font size}
}
\value{
A plotly object
}
\description{
Change plotly font to Arial
}
\examples{
library(plotly)
data(mtcars)

p <- plot_ly(mtcars, x = ~mpg, y = ~wt, color = ~cyl,
             type = 'scatter', mode = 'markers')
p \%>\% arial
p \%>\% arial(18)
}
