% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easylabel.R
\name{easylabel}
\alias{easylabel}
\title{Interactive scatter plot labels}
\usage{
easylabel(
  data,
  x,
  y,
  col,
  labs = NULL,
  scheme = NULL,
  xlab = x,
  ylab = y,
  startLabels = NULL,
  xlim = NULL,
  ylim = NULL,
  symbols = c("circle", "diamond-open"),
  showOutliers = TRUE,
  width = 800,
  height = 600,
  showgrid = FALSE,
  zeroline = TRUE,
  hline = NULL,
  vline = NULL,
  markerSize = 8,
  alpha = 1,
  markerOutline = list(width = 0.5, color = "white"),
  marker = list(size = markerSize, opacity = alpha, line = markerOutline),
  custom_annotation = NULL
)
}
\arguments{
\item{data}{Dataset to use for plot.}

\item{x}{specifies column of x coordinates in \code{data}.}

\item{y}{specifies column of y coordinates in \code{data}.}

\item{col}{specifies the column (ideally a factor) in \code{data} with which
to colour points.}

\item{labs}{specifies the column in \code{data} with label names for points.
If NULL defaults to use \code{rownames(data)}.}

\item{scheme}{vector of colours for points.}

\item{xlab}{x axis title.}

\item{ylab}{y axis title.}

\item{startLabels}{vector of initial points to be labelled.}

\item{xlim}{the x limits (x1, x2) of the plot.}

\item{ylim}{the y limits of the plot.}

\item{symbols}{passed to plotly to specify symbols for normal points and
outliers.}

\item{showOutliers}{Logical whether to show outliers on the margins of the
plot.}
}
\description{
Interactive labelling of scatter plots using shiny and plotly interface.
}
\details{
Instructions:
\enumerate{
\item Hover over and click on/off genes which you want to label.
\item When you have selected all your chosen genes, then drag gene names to
move label positions.
\item You can move the legend as well.
\item Switch to SVG when finalised (only do this at last moment as otherwise
editing is very slow).
\item Press camera button in modebar to save image as SVG.
}
}
