% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqplot.R
\name{gg_qqplot}
\alias{gg_qqplot}
\title{Log QQ p-value plot (ggplot2)}
\usage{
gg_qqplot(df, scheme = c("darkgrey", "royalblue"))
}
\arguments{
\item{df}{A dataframe generated by \code{\link[=qqplot]{qqplot()}}}

\item{scheme}{Vector of 2 colours for plotting non-significant and
significant SNPs}
}
\value{
A ggplot2 graphics plot object
}
\description{
Produces a QQ plot via ggplot2. Requires a dataframe generated by
\code{\link[=qqplot]{qqplot()}}.
}
