% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadlabel.R
\name{loadlabel}
\alias{loadlabel}
\title{Load easylabel saved state}
\usage{
loadlabel(object, data = NULL, ...)
}
\arguments{
\item{object}{Either a character value specifying an rds file to be loaded,
or an object of class 'easylab'.}

\item{data}{Dataset (data.frame or data.table) to use for plot. If not
specified, the function will first use embedded data included in \code{object};
if no data is embedded it will attempt to load the original data object as
it was specified in the original call to \code{\link[=easylabel]{easylabel()}}.}

\item{...}{Additional arguments passed to \code{\link[=easylabel]{easylabel()}}. Can be used to
overwrite original arguments.}
}
\value{
By default no return value. Calls \code{\link[=easylabel]{easylabel()}} which will open a
shiny interactive session based on previous settings stored in an 'easylab'
class object or rds file of such an object.
}
\description{
Loads a file or easylab object containing saved state with specified labels
and their positions.
}
