% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqplot.R
\name{qqplot2}
\alias{qqplot2}
\title{Log QQ p-value plot (2nd stage)}
\usage{
qqplot2(df, scheme = c("darkgrey", "royalblue"), ...)
}
\arguments{
\item{df}{A dataframe generated by \code{\link[=qqplot]{qqplot()}}}

\item{scheme}{Vector of 2 colours for plotting non-significant and
significant SNPs}

\item{...}{Optional plotting arguments passed to \code{\link[=plot]{plot()}}}
}
\value{
No return value. Produces a base graphics plot.
}
\description{
Second stage plotting function which accepts dataframe generated by
\code{\link[=qqplot]{qqplot()}}. This can be used to avoid repeating computation of the QQ plot
values.
}
