% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadlabel.R
\name{summary.easylab}
\alias{summary.easylab}
\title{Summarise easylabel saved state object}
\usage{
\method{summary}{easylab}(object, ...)
}
\arguments{
\item{object}{Object of class 'easylab'}

\item{...}{Optional arguments for compatibility}
}
\value{
No return value. Prints information on the original easylabel call
and saved labels.
}
\description{
Prints summary information on an easylabel saved state object.
}
