% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{volc1}
\alias{volc1}
\title{Example volcano data for vignette}
\format{
Data frame with 6 rows and 6 variables
}
\usage{
volc1
}
\description{
Example DESeq2 volcano data for vignette
}
\keyword{datasets}
