% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easylabel.R
\name{volcanoplot}
\alias{volcanoplot}
\title{Interactive volcano plot labels}
\usage{
volcanoplot(
  data,
  x = NULL,
  y = NULL,
  fdrcutoff = 0.05,
  fccut = 0,
  scheme = c("darkgrey", "red"),
  show_counts = TRUE,
  useQ = FALSE,
  ...
)
}
\arguments{
\item{data}{The dataset for the plot. Automatically attempts to recognises
DESeq2 and limma objects.}

\item{x}{column containing log fold change. For DESeq2 and limma objects
this is automatically set.}

\item{y}{column containing p values. For DESeq2 and limma objects this is
automatically set.}

\item{fdrcutoff}{Cut-off for FDR significance. Defaults to FDR < 0.05}

\item{fccut}{Log fold change cut-off. Optional, where 0 means not applied.}

\item{scheme}{Colour scheme. If no fold change cut-off is set, 2 colours
need to be specified. With a fold change cut-off, 3 colours are required.}

\item{show_counts}{Logical whether to show legend with number of
differentially expressed genes.}

\item{useQ}{Logical whether to convert nominal P values to q values.
Requires the qvalue Bioconductor package.}

\item{...}{other arguments passed to \code{\link[=easylabel]{easylabel()}}.}
}
\description{
Interactive labelling of volcano plots using shiny/plotly interface.
}
\seealso{
\code{\link[=easylabel]{easylabel()}}
}
