% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_labels.R
\name{add_labels}
\alias{add_labels}
\title{Add labels to a plotly scatter plot}
\usage{
add_labels(p, labs)
}
\arguments{
\item{p}{A plotly scatter plot object}

\item{labs}{Character vector of labels to match}
}
\value{
A plotly plot with added labels
}
\description{
Add labels to a plotly scatter plot
}
