% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easylabel.R
\name{easylabel}
\alias{easylabel}
\title{Interactive scatter plot labels}
\usage{
easylabel(
  data,
  x,
  y,
  col,
  labs = NULL,
  scheme = NULL,
  xlab = x,
  ylab = y,
  startLabels = NULL,
  xlim = NULL,
  ylim = NULL,
  symbols = c("circle", "diamond-open"),
  showOutliers = TRUE,
  width = 800,
  height = 600,
  showgrid = FALSE,
  zeroline = TRUE,
  hline = NULL,
  vline = NULL,
  markerSize = 8,
  alpha = 1,
  markerOutline = list(width = 0.5, color = "white"),
  marker = list(size = markerSize, opacity = alpha, line = markerOutline),
  custom_annotation = NULL
)
}
\arguments{
\item{data}{Dataset to use for plot.}

\item{x}{specifies column of x coordinates in \code{data}.}

\item{y}{specifies column of y coordinates in \code{data}.}

\item{col}{specifies the column (ideally a factor) in \code{data} with which
to colour points.}

\item{labs}{specifies the column in \code{data} with label names for points.
If \code{NULL} defaults to use \code{rownames(data)}.}

\item{scheme}{vector of colours for points.}

\item{xlab}{x axis title.}

\item{ylab}{y axis title.}

\item{startLabels}{vector of initial labels.}

\item{xlim}{the x limits (x1, x2) of the plot.}

\item{ylim}{the y limits of the plot.}

\item{symbols}{passed to plotly to specify symbols for normal points and
outliers.}

\item{showOutliers}{Logical whether to show outliers on the margins of the
plot.}

\item{width}{Width of the plot in pixels.}

\item{height}{Height of the plot in pixels.}

\item{showgrid}{Logical whether to show plotly gridlines.}

\item{zeroline}{Logical whether to show lines at x=0 and y=0.}

\item{hline}{Adds horizontal lines at values of y.}

\item{vline}{Adds vertical lines at values of x.}

\item{markerSize}{Size of markers as per plotly.}

\item{alpha}{Alpha value for transparency of points.}

\item{markerOutline}{List of plotly arguments to define marker outlines.}

\item{marker}{List of arguments to control plotly markers.}

\item{custom_annotation}{List of annotations to be added via \code{\link[plotly:layout]{plotly::layout()}}.}
}
\description{
Interactive labelling of scatter plots using shiny and plotly interface.
}
\details{
Instructions:
\itemize{
\item Hover over and click on/off genes which you want to label.
\item When you have selected all your chosen genes, then drag gene names to
move label positions.
\item You can move the legend as well.
\item Switch to SVG when finalised (only do this at last moment as otherwise
editing is very slow).
\item Press camera button in modebar to save image as SVG.
\item The Table tab shows a table view of the dataset to help with annotation.
}
}
\seealso{
\code{\link[=plot_ly]{plot_ly()}}
}
